==> Semster A -Clear information and statistical tests/הוספת זיהום לדאטה/1.cleanAndCreateDFS.ipynb <==
Imports

import numpy as np
import pandas as pd

read dfs

pm25_df = pd.read_excel("pm25.xlsx")
pm10_df = pd.read_excel("pm10.xlsx")
Nox_df = pd.read_excel("Nox.xlsx")
So2_df = pd.read_excel("So2.xlsx")

dfs=[pm25_df,pm10_df,Nox_df,So2_df ]
dfs[0]

row 0 is all nan
so we slice the first row and last 10 are non relevnt


for i in range (len(dfs)):
    dfs[i] = dfs[i][1:]
    dfs[i] = dfs[i][:-10]

Names the column by the cities in the first new row

for i in range (len(dfs)):
    cols_names = dfs[i][:1]
    dfs[i] = dfs[i][1:]
    dfs[i].columns = np.array( cols_names)[0]


Now two first Rows are type of pollution and std (i think)

for i in range (len(dfs)):
    dfs[i] = dfs[i][2:]

Chande the column date to date type

for i in range (len(dfs)):
    #no need it the string 24:00 and other stiring
    dfs[i][dfs[i].columns[0]]= dfs[i][dfs[i].columns[0]].replace("24:00","" ,regex=True)
    dfs[i]= dfs[i].replace("NoData", np.NaN, regex=True)
    dfs[i]= dfs[i].replace("&lt;Samp", np.NaN, regex=True)
    dfs[i]= dfs[i].replace("Down", np.NaN, regex=True)
    dfs[i]= dfs[i].replace("InVld", np.NaN, regex=True)
    dfs[i]= dfs[i].replace("Calib", np.NaN, regex=True)
    dfs[i]= dfs[i].replace("_", np.NaN, regex=True)

    dfs[i][dfs[i].columns[0]] = pd.to_datetime( dfs[i][dfs[i].columns[0]])
    for j in range(1,len(dfs[i].columns)):
        dfs[i][dfs[i].columns[j]] = dfs[i][dfs[i].columns[j]].astype("float64")


save in the orignal dfs

pm25_df = dfs[0] 
pm10_df = dfs[1]
Nox_df = dfs[2]
So2_df = dfs[3]



pm25_df.to_excel("pm25_cleaner_NEW_WITH_NO_DROP_CITIES.xlsx")
pm10_df.to_excel("pm10_cleaner_NEW_WITH_NO_DROP_CITIES.xlsx")
Nox_df.to_excel("Nox_cleaner_NEW_WITH_NO_DROP_CITIES.xlsx")
So2_df.to_excel("So2_cleaner_NEW_WITH_NO_DROP_CITIES.xlsx")

Change df to date,city and value

for i in range (len(dfs)):
    dfs[i] = pd.melt(dfs[i], 'תאריך ושעה')
    dfs[i] = dfs[i].sort_values("תאריך ושעה")
    dfs[i].reset_index(drop=True,inplace=True)
    dfs[i].columns = ["Date","City","Value"]

change column name by pollutions

dfs[0].columns = ["Date","City","PM25"]
dfs[1].columns = ["Date","City","PM10"]
dfs[2].columns = ["Date","City","Nox"]
dfs[3].columns = ["Date","City","So2"]



# name citiys 

citiys = [ word.split(",")[0] for i in range (4) for word in dfs[i].City.unique() ]
citiys = list(set(citiys))

#change cities name for more comfortable
for i in range(4):
    for j in range(len(citiys)):
        dfs[i].City[dfs[i].City.str.contains(citiys[j])] = citiys[j]

for i in range(4):
    tmp = dfs[i]
    tmp2 = tmp.groupby(["Date",'City'], sort=False)[tmp.columns[-1]].max()
    tmp3=tmp2.reset_index()
    dfs[i] = tmp3

merge dfs

df_m=dfs[0]
df_m=df_m.merge(right=dfs[1],how="inner",on=["Date","City"])
df_m=df_m.merge(right=dfs[2],how="inner",on=["Date","City"])
df_m=df_m.merge(right=dfs[3],how="inner",on=["Date","City"])
df_m.to_excel('merag.xlsx')

merage all dfs into one and save it without null

print(df_m.shape)
df_m_NoNan=df_m.dropna(thresh=6)
df_m_NoNan.shape

Check how many dates have been lost

print("Number of unique dates: with null",df_m.Date.nunique())
print("Number of unique dates: with without null",df_m_NoNan.Date.nunique())

Save merage without null

df_m_NoNan.to_excel("merageWithoutNull.xlsx")


df_m_NoNan["Date"].value_counts().nlargest(1000)

# dfs[0].isna().sum()/len(dfs[0])*100
# pm25_df = dfs[0] 
# pm10_df = dfs[1]
# Nox_df = dfs[2]
# So2_df = dfs[3]
# pm25_df

df_m_NoNan.head()

df_m_NoNan.City.unique()
