==> Semster A -Clear information and statistical tests/הוספת זיהום לדאטה/2.remove duplicat cols.ipynb <==
import numpy as np
import pandas as pd

pm25_df = pd.read_excel("pm25_cleaner_NEW_WITH_NO_DROP_CITIES.xlsx",date_parser={"תאריך ושעה"})
pm10_df = pd.read_excel("pm10_cleaner_NEW_WITH_NO_DROP_CITIES.xlsx",date_parser={"תאריך ושעה"})
Nox_df = pd.read_excel("Nox_cleaner_NEW_WITH_NO_DROP_CITIES.xlsx",date_parser={"תאריך ושעה"})
So2_df = pd.read_excel("So2_cleaner_NEW_WITH_NO_DROP_CITIES.xlsx",date_parser={"תאריך ושעה"})



pm25_df.head(1)

cols = [word.split(",")[0] for word in pm25_df.columns]
pm25_df.columns = cols

cols = [word.split(",")[0] for word in pm10_df.columns]
pm10_df.columns = cols

cols = [word.split(",")[0] for word in Nox_df.columns]
Nox_df.columns = cols

cols = [word.split(",")[0] for word in So2_df.columns]
So2_df.columns = cols

len(So2_df.columns)

So2_df=pd.concat([So2_df[col].sum(axis=1).rename(col) if len(So2_df[col].shape)==2 else So2_df[col] for col in So2_df.columns.unique()],axis=1)
pm10_df=pd.concat([pm10_df[col].sum(axis=1).rename(col) if len(pm10_df[col].shape)==2 else pm10_df[col] for col in pm10_df.columns.unique()],axis=1)
pm25_df=pd.concat([pm25_df[col].sum(axis=1).rename(col) if len(pm25_df[col].shape)==2 else pm25_df[col] for col in pm25_df.columns.unique()],axis=1)
Nox_df=pd.concat([Nox_df[col].sum(axis=1).rename(col) if len(Nox_df[col].shape)==2 else Nox_df[col] for col in Nox_df.columns.unique()],axis=1)


pm25_df.to_excel("pm25_cleaner_NEW_WITH_NO_DROP_CITIES.xlsx")
pm10_df.to_excel("pm10_cleaner_NEW_WITH_NO_DROP_CITIES.xlsx")
So2_df.to_excel("So2_cleaner_NEW_WITH_NO_DROP_CITIES.xlsx")
Nox_df.to_excel("Nox_cleaner_NEW_WITH_NO_DROP_CITIES.xlsx")

Nox_df
