==> Semster A -Clear information and statistical tests/הוספת זיהום לדאטה/3.transpose dfs.ipynb <==
import numpy as np
import pandas as pd

pm25_df = pd.read_excel("pm25_cleaner_NEW_WITH_NO_DROP_CITIES.xlsx",date_parser={"תאריך ושעה"})
pm10_df = pd.read_excel("pm10_cleaner_NEW_WITH_NO_DROP_CITIES.xlsx",date_parser={"תאריך ושעה"})
Nox_df = pd.read_excel("Nox_cleaner_NEW_WITH_NO_DROP_CITIES.xlsx",date_parser={"תאריך ושעה"})
So2_df = pd.read_excel("So2_cleaner_NEW_WITH_NO_DROP_CITIES.xlsx",date_parser={"תאריך ושעה"})


So2_df

pm25_df.drop(pm25_df.columns[0],axis=1,inplace=True)
pm10_df.drop(pm10_df.columns[0],axis=1,inplace=True)
Nox_df.drop(Nox_df.columns[0],axis=1,inplace=True)
So2_df.drop(So2_df.columns[0],axis=1,inplace=True)

So2_df.drop("Unnamed: 0.1",axis=1,inplace=True)
pm10_df.drop("Unnamed: 0.1",axis=1,inplace=True)
Nox_df.drop("Unnamed: 0.1",axis=1,inplace=True)
pm25_df.drop("Unnamed: 0.1",axis=1,inplace=True)

def changeDF(df):
    df = pd.melt(df, 'תאריך ושעה')
    df = df.sort_values("תאריך ושעה")
    df.reset_index(drop=True,inplace=True)
    df.columns = ["Date","City","Value"]
    return df
    
pm25_df = changeDF(pm25_df)
pm10_df = changeDF(pm10_df)
Nox_df = changeDF(Nox_df)
So2_df = changeDF(So2_df)


So2_df

pm25_df.to_excel("pm25_T_NEW_WITH_NO_DROP_CITIES.xlsx",sheet_name = "pm25_T_NEW_WITH_NO_DROP_CITIES", index=False)
pm10_df.to_excel("pm10_T_NEW_WITH_NO_DROP_CITIES.xlsx",sheet_name = "pm10_T_NEW_WITH_NO_DROP_CITIES", index=False)
So2_df.to_excel("So2_T_NEW_WITH_NO_DROP_CITIES.xlsx",sheet_name = "So2_T_NEW_WITH_NO_DROP_CITIES", index=False)
Nox_df.to_excel("Nox_T_NEW_WITH_NO_DROP_CITIES.xlsx",sheet_name = "Nox_T_NEW_WITH_NO_DROP_CITIES", index=False)
