==> Semster A -Clear information and statistical tests/הוספת זיהום לדאטה/4.merage_df_T.ipynb <==
import pandas as pd

pm25_df = pd.read_excel("pm25_T_NEW_WITH_NO_DROP_CITIES.xlsx",date_parser={"Date"})
pm10_df = pd.read_excel("pm10_T_NEW_WITH_NO_DROP_CITIES.xlsx",date_parser={"Date"})
Nox_df = pd.read_excel("Nox_T_NEW_WITH_NO_DROP_CITIES.xlsx",date_parser={"Date"})
So2_df = pd.read_excel("So2_T_NEW_WITH_NO_DROP_CITIES.xlsx",date_parser={"Date"})
So2_df.info()


So2_df.head()

merage_df = pm10_df.merge(right=pm25_df,on=["Date","City"],how="outer")
merage_df = merage_df.merge(right=Nox_df,on=["Date","City"],how="outer")
merage_df = merage_df.merge(right=So2_df,on=["Date","City"],how="outer")
merage_df

merage_df.columns=['Date',"City",'pm10','pm2.5','nox','so2']
merage_df.sort_values(["Date"],inplace=True)
merage_df

print(len(merage_df))
merage_df=merage_df.dropna(thresh=3)
cities = merage_df.City.unique()
len(merage_df)

print(cities)

from geopy.geocoders import Nominatim
loc = Nominatim(user_agent="GetLoc")
coo = {}
for city in cities:
    try:
        getLoc = loc.geocode(city)
        if(getLoc is not type(None)):
            coo[city] = [getLoc.latitude,getLoc.longitude]
    except :
        continue





print(len(cities))
len(coo)
for c in cities:
    if c not in coo:
        print(c)

coo['מועצה תעשייתית נאות חובב'] = [31.135,34.791]
coo['קיריית טבעון'] = [32.715603,35.126578]
coo['חיפה וקריות'] = [32.841592,35.071278]
coo['ארז'] =[31.56028,34.566483]
coo['מגל'] =[32.385837,35.03366]
coo['ערד'] =[31.258928,35.213893]
coo['גן דרום'] = [31.804468,34.700824]

coo['ברקאי'] = [32.474819141691206,35.0275950342368]
coo['חריש'] = [32.462194973207886,35.045251620627795]
coo['קיסריה'] = [32.50712927639148,34.91244095275816]
coo['אלעד'] = [32.05089076091389,34.95718705819683]
coo['אשלים'] = [30.963449095763554,34.70081419534035]

merage_df['latitude'] = 0
merage_df['longitude'] = 0

for key,val in coo.items():
    merage_df.latitude[merage_df["City"]==key] = val[0]
    merage_df.longitude[merage_df["City"]==key] = val[1]

merage_df.reset_index(drop=True,inplace=True)

df=pd.DataFrame.from_dict(coo)
df=df.transpose()
df.columns=['lat','lon']

import folium

lat = df["lat"][0]
lon = df["lon"][0]

map = folium.Map(location=[lat, lon], zoom_start=3)
for i in range (len(df.lat)):
    folium.Marker([df["lat"][i], df['lon'][i]],popup= folium.Popup(df.index[i])).add_to(map)
    folium.Circle([df["lat"][i], df['lon'][i]]).add_to(map)

map

# from geopy import distance

# cities_in_corr = {}
# def inRadius(CityA,lanA,lonA,CityB,lanB,lonB):
#     center_point = [{lanA, lonA}]
#     test_point = [{lanB, lonB}]
#     radius = 5 # in kilometer
#     center_point_tuple = tuple(center_point[0]) # (-7.7940023, 110.3656535)
#     test_point_tuple = tuple(test_point[0]) # (-7.79457, 110.36563)

#     dis = distance.distance(center_point_tuple, test_point_tuple).km
#     # print("Distance: {}".format(dis)) # Distance: 0.0628380925748918

#     if dis <= radius:
#         if CityA not in cities_in_corr.keys():
#             cities_in_corr[CityA] = [CityB]
#         else:
#             cities_in_corr[CityA].append(CityB)
            
#         if CityB not in cities_in_corr.keys():
#             cities_in_corr[CityB] = [CityA]
#         else:
#             cities_in_corr[CityB].append(CityA)
#         # print("{} point is inside the {} km radius from {} coordinate".format(CityB, radius, CityA))
#     # else:
#         # print("{} point is outside the {} km radius from {} coordinate".format(CityB, radius, CityA))

# for i in range (len(df.index)):
#     for j in range(i+1,len(df.index)):
#         inRadius(df.index[i],df.lat[i],df.lon[i],df.index[j],df.lat[j],df.lon[j])


# print("total in radius 5 km",len(cities_in_corr))
# print("total citiy",len(cities))
# cities_in_corr

merage_df.to_excel("merge_NEW_WITH_NO_DROP_CITIES.xlsx")

# from math import sin, cos, sqrt, atan2

# R = 6373.0

# lat1 = 52.2296756
# lon1 = 21.0122287
# lat2 = 52.406374
# lon2 = 16.9251681

# dlon = lon2 - lon1
# dlat = lat2 - lat1
# a = (sin(dlat/2))**2 + cos(lat1) * cos(lat2) * (sin(dlon/2))**2
# c = 2 * atan2(sqrt(a), sqrt(1-a))
# distance = R * c

# print ("Result", distance)
# print( "Should be", 278.546)
