==> Semster A -Clear information and statistical tests/הוספת זיהום לדאטה/7.merge_SitePollution_with_SiteInfo.ipynb <==
import pandas as pd

df_site_info=pd.read_excel("test.xlsx")
df_site_poll=pd.read_excel("polluions_sits_df.xlsx")

df_site_poll.drop('Unnamed: 0',axis=1,inplace=True)
df_site_info.drop('Unnamed: 0',axis=1,inplace=True)

df_site_info.head(1)

df_site_poll.head(1)

print(len(df_site_info))
df_site_info=df_site_info.merge(right=df_site_poll,on=["Date","Site_Name"])
print(len(df_site_info))
df_site_info.head()

df_site_info.to_excel("df_site_info.xlsx",sheet_name="df_site_info")

df = pd.melt(df, 'תאריך ושעה')
df = df.sort_values("תאריך ושעה")
df.reset_index(drop=True,inplace=True)
df.columns = ["Date","City","Value"]
